Added reschedule that doesn't reset reps, lapses.

+++ 2089200096/__init__.py
@@ -10,7 +10,7 @@ from anki.utils import ids2str, intTime
 
 
 # Modified copy of `Scheduler.reschedCards' that doesn't change `ease' factor.
-def reschedCards(scheduler, cids):
+def reschedCards(scheduler, cids, isFullReset=False):
     "Put cards in review queue with a new interval of 1 day."
     d = []
     due = scheduler.today
@@ -19,15 +19,25 @@ def reschedCards(scheduler, cids):
     for id in cids:
         d.append((due, usn, mod, id,))
     scheduler.remFromDyn(cids)
-    scheduler.col.db.executemany(
-        f"""
-update cards set odue=0,reps=0,lapses=0,ivl=1,
-due=?,usn=?,mod=? where id=?""",
-        d,
-    )
+
+    if isFullReset:
+        scheduler.col.db.executemany(
+            f"""
+            update cards set odue=0,reps=0,lapses=0,ivl=1,
+            due=?,usn=?,mod=? where id=?""",
+            d,
+        )
+    else:
+        scheduler.col.db.executemany(
+            f"""
+            update cards set ivl=1,
+            due=?,usn=?,mod=? where id=?""",
+            d,
+        )
+
     scheduler.col.log(cids)
 
-def onResetCards(browser):
+def onResetCards(browser, isFullReset=False):
     cids = browser.selectedCards()
     if not cids:
         showInfo("No cards selected.")
@@ -37,10 +47,12 @@ def onResetCards(browser):
 
     mw.col.modSchema(check=True)
     mw.progress.start(immediate=True)
-    mw.col.db.execute("delete from revlog where cid in " + ids2str(cids))
+    if isFullReset:
+        mw.col.db.execute("delete from revlog where cid in " + ids2str(cids))
     
-    # Reschedule cards for today and reset `reps' and `lapses' counters.
-    reschedCards(mw.col.sched, cids)
+    # Reschedule cards for today and, when `isFullReset' is True,
+    # reset `reps' and `lapses' counters.
+    reschedCards(mw.col.sched, cids, isFullReset)
     
     mw.col.setMod()
     mw.col.save()
@@ -52,11 +64,17 @@ def onResetCards(browser):
     showInfo("Reset %d cards" % len(cids))
 
 def onMenuSetup(browser):
-    act = QAction(browser)
-    act.setText("Reset Cards")
     mn = browser.form.menu_Cards
     mn.addSeparator()
+    
+    act = QAction(browser)
+    act.setText("Reset Cards")
     mn.addAction(act)
-    act.triggered.connect(lambda b=browser: onResetCards(browser))
+    act.triggered.connect(lambda b=browser: onResetCards(browser, True))
+    
+    act = QAction(browser)
+    act.setText("Reschedule Cards for Today")
+    mn.addAction(act)
+    act.triggered.connect(lambda b=browser: onResetCards(browser, False))
 
 addHook("browser.setupMenus", onMenuSetup)
